+++ Services/Aad/V2/Model/ListConnectionNumberDataList.cs
@@ -20,7 +20,7 @@ namespace HuaweiCloud.SDK.Aad.V2.Model
         /// 时间戳毫秒值
         /// </summary>
         [JsonProperty("time", NullValueHandling = NullValueHandling.Ignore)]
-        public int? Time { get; set; }
+        public long? Time { get; set; }
 
         /// <summary>
         /// 连接数