+++ Services/GaussDBforopenGauss/V3/GaussDBforopenGaussAsyncClient.cs
@@ -5532,7 +5532,6 @@ namespace HuaweiCloud.SDK.GaussDBforopenGauss.V3
         public async Task<ListLtsConfigsResponse> ListLtsConfigsAsync(ListLtsConfigsRequest listLtsConfigsRequest)
         {
             var urlParam = new Dictionary<string, string>();
-            if (StringUtils.TryConvertToNonEmptyString(listLtsConfigsRequest.InstanceId, out var valueOfInstanceId)) urlParam.Add("instance_id", valueOfInstanceId);
             var urlPath = HttpUtils.AddUrlPath("/v3/{project_id}/instances/logs/lts-config", urlParam);
             var request = HttpUtils.InitSdkRequest(urlPath, "application/json", listLtsConfigsRequest);
             var response = await DoHttpRequestAsync("GET", request);
@@ -5542,7 +5541,6 @@ namespace HuaweiCloud.SDK.GaussDBforopenGauss.V3
         public AsyncInvoker<ListLtsConfigsResponse> ListLtsConfigsAsyncInvoker(ListLtsConfigsRequest listLtsConfigsRequest)
         {
             var urlParam = new Dictionary<string, string>();
-            if (StringUtils.TryConvertToNonEmptyString(listLtsConfigsRequest.InstanceId, out var valueOfInstanceId)) urlParam.Add("instance_id", valueOfInstanceId);
             var urlPath = HttpUtils.AddUrlPath("/v3/{project_id}/instances/logs/lts-config", urlParam);
             var request = HttpUtils.InitSdkRequest(urlPath, "application/json", listLtsConfigsRequest);
             return new AsyncInvoker<ListLtsConfigsResponse>(this, "GET", request, JsonUtils.DeSerialize<ListLtsConfigsResponse>);