+++ Services/GaussDBforopenGauss/V3/Model/TableVolumeResult.cs
@@ -53,13 +53,13 @@ namespace HuaweiCloud.SDK.GaussDBforopenGauss.V3.Model
         public string DatabaseName { get; set; }
 
         /// <summary>
-        /// **参数解释**: 表或者索引是否具有分区表的性质。 **取值范围**: 不涉及。 
+        /// **参数解释**: 表或者索引是否具有分区表的性质。 **取值范围**: - true：具有分区表的性质。 - false：不具有分区表的性质。 
         /// </summary>
         [JsonProperty("is_part_type", NullValueHandling = NullValueHandling.Ignore)]
         public bool? IsPartType { get; set; }
 
         /// <summary>
-        /// **参数解释**: 是否包含hash分区列信息。 **取值范围**: 不涉及。 
+        /// **参数解释**: 是否包含hash分区列信息。 **取值范围**: - true：包含hash分区列信息。 - false：不包含hash分区列信息。 
         /// </summary>
         [JsonProperty("is_hash_cluster_key", NullValueHandling = NullValueHandling.Ignore)]
         public bool? IsHashClusterKey { get; set; }
@@ -113,7 +113,7 @@ namespace HuaweiCloud.SDK.GaussDBforopenGauss.V3.Model
         public string SkewRatio { get; set; }
 
         /// <summary>
-        /// **参数解释**: 表分布标准方差（在表大小一定的情况下，该值越大表明表的整体分布情况越倾斜）。。 **取值范围**: 不涉及。 
+        /// **参数解释**: 表分布标准方差（在表大小一定的情况下，该值越大表明表的整体分布情况越倾斜）。 **取值范围**: 不涉及。 
         /// </summary>
         [JsonProperty("skew_stddev", NullValueHandling = NullValueHandling.Ignore)]
         public string SkewStddev { get; set; }